+++ components/clientComponents/clientSubText.tsx
@@ -1,0 +1,26 @@
+import styled from "styled-components";
+
+import { clientFonts } from "@/styles/theme";
+
+interface ClientSubTextInterface {
+  center?: boolean;
+  children: string | string[];
+}
+
+const ClientSubTextWrap = styled.span`
+  width: 100%;
+  font-weight: 500;
+  display: block;
+  font-size: ${clientFonts.sm};
+  text-align: ${(props: { center: boolean }) =>
+    props.center ? "center" : null};
+`;
+
+const ClientSubText = ({
+  children,
+  center = false,
+}: ClientSubTextInterface) => {
+  return <ClientSubTextWrap center={center}>{children}</ClientSubTextWrap>;
+};
+
+export default ClientSubText;
