+++ components/clientComponents/index.ts
@@ -2,5 +2,16 @@ import HistoryBackButton from "./historyBackButton";
 import Layout from "./layout";
 import ClientButton from "./clientButton";
 import IntroLabel from "./introLabel";
+import BottomButtonLayout from "./bottomButtonLayout";
+import ClientText from "./clientText";
+import ClientSubText from "./clientSubText";
 
-export { HistoryBackButton, Layout, ClientButton, IntroLabel };
+export {
+  HistoryBackButton,
+  Layout,
+  ClientButton,
+  IntroLabel,
+  BottomButtonLayout,
+  ClientText,
+  ClientSubText,
+};
